Luke is fixing things
Fixing the Main activity and vacations list, We can now get friends
vacations. Redone a bunch of code to make it work better.

things to do by tonight:
Linking view vacation to the main activity list
Set up the array adapter to display vacation title, place and Creator
Need to redesign the mainactivity to fit in the with the blue design
style